+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 ### Unreleased
 
+#### BREAKING
+
+Note: though this are technically breaking changes, they are unlikely to break
+most setups. They are only a problem if you used scripts that relied on
+`swww query` output format, or if you relied on the `swww-daemon` socket file
+name.
+
+  * `swww query` now returns information in the following format:
+  `<namespace>: OUTPUT: SIZE, scale: SCALE, currently displaying: IMAGE_OR_COLOR`
+  * we are creating the socket in a slightly different location now:
+  `${XDG_RUNTIME_DIR}/${WAYLAND_DISPLAY}-swww-daemon.<specified namespace>.socket`
+
+#### Additions
+
+  * every command (except `clear-cache`) now accepts a `-n|--namespace`
+  argument. This namespace will then be appended to `swww-daemon` to make up the
+  wayland namespace of the daemon's layer surface.
+  * as a convenience, every command (except `clear-cache`) also accepts a
+  `-a|--all` flag, that lets the user send commands to all namespaces at once.
 
 ### 0.10.3